Add styling and permission check for JobPost and Department
Add permission check for specific pages like delete for admin
Add messages to the user for add, edit, and delete for JobPost and Department
Add dollar value format for the JobPost (stored in the database as cents CAD) and change model for JobPost to use a int datatype for salary
Styling the JobPost and Department feature
Add code for future development of a different feature (Job Apply) in the JobPost details page

+++ Models/JobPost.cs
@@ -30,8 +30,8 @@ namespace HSNHospitalProject.Models
         //JobPost has a filled flag (if the job is taken or not)
         public bool jobPostFilled { get; set; }
 
-        //JobPost has a salary rate
-        public float jobPostSalary { get; set; }
+        //JobPost has a salary rate ($CAD and value is whole number by multiplying by 100) 
+        public int jobPostSalary { get; set; }
 
         //JobPost has a posted date
         public DateTime jobPostPostedDate { get; set; }
